remove useless lines from codebase
