+++ pkg/transport-discovery/api/api.go
@@ -104,7 +104,6 @@ func New(log logrus.FieldLogger, s store.Store, nonceStore httpauth.NonceStore,
 	r.Get("/stats", api.stats)
 	r.Get("/all-transports", api.getAllTransports)
 	r.Delete("/transports/deregister", api.deregisterTransport)
-	//	r.Post("/statuses", func(w http.ResponseWriter, r *http.Request) {
 	r.Post("/statuses", func(w http.ResponseWriter, _ *http.Request) {
 		w.WriteHeader(http.StatusGone)
 	})
