+++ pkg/transport-discovery/api/endpoints.go
@@ -239,7 +239,6 @@ func (api *API) logger(r *http.Request) logrus.FieldLogger {
 	return httputil.GetLogger(r)
 }
 
-// stats endpoint for monitoring
 func (api *API) stats(w http.ResponseWriter, r *http.Request) {
 	if ms, ok := api.store.(*store.MemoryStore); ok {
 		stats := ms.GetStats()
