+++ pkg/transport-discovery/store/postgres_store.go
@@ -2,15 +2,12 @@ package store
 
 import (
 	"context"
-	"time"
 
 	"github.com/google/uuid"
 	"gorm.io/gorm"
 
-	"github.com/skycoin/skywire/pkg/skywire-utilities/pkg/cipher"
 	"github.com/skycoin/skywire/pkg/skywire-utilities/pkg/logging"
 	"github.com/skycoin/skywire/pkg/transport"
-	types "github.com/skycoin/skywire/pkg/transport/types"
 )
 
 type postgresStore struct {
@@ -18,22 +15,6 @@ type postgresStore struct {
 	client *gorm.DB
 	cache  map[string]int64
 	closeC chan struct{}
-}
-
-// NewPostgresStore creates new sransports postgres store.
-func NewPostgresStore(logger *logging.Logger, cl *gorm.DB) (TransportStore, error) {
-	// automigrate
-	if err := cl.AutoMigrate(Transport{}); err != nil {
-		logger.Warn("failed to complete automigrate process")
-	}
-
-	s := &postgresStore{
-		log:    logger,
-		client: cl,
-		cache:  make(map[string]int64),
-		closeC: make(chan struct{}),
-	}
-	return s, nil
 }
 
 func (s *postgresStore) RegisterTransport(_ context.Context, sEntry *transport.SignedEntry) error {
@@ -53,111 +34,6 @@ func (s *postgresStore) DeregisterTransport(ctx context.Context, id uuid.UUID) e
 	return s.client.Where("transport_id = ?", id).Delete(&Transport{}).Error
 }
 
-func (s *postgresStore) GetTransportByID(_ context.Context, id uuid.UUID) (*transport.Entry, error) {
-	var tpRecord Transport
-	if err := s.client.Where("transport_id = ?", id).First(&tpRecord).Error; err != nil {
-		return nil, ErrTransportNotFound
-	}
-
-	entry, err := makeEntry(tpRecord)
-	if err != nil {
-		return nil, err
-	}
-
-	return &entry, nil
-}
-
-func (s *postgresStore) GetTransportsByEdge(_ context.Context, pk cipher.PubKey) ([]*transport.Entry, error) {
-	var tpRecords []Transport
-	if err := s.client.Where("edge_a = ? OR edge_b = ?", pk.Hex(), pk.Hex()).Find(&tpRecords).Error; err != nil {
-		return nil, ErrTransportNotFound
-	}
-
-	var entries []*transport.Entry
-
-	for _, tpRecord := range tpRecords {
-		entry, err := makeEntry(tpRecord)
-		if err != nil {
-			return nil, err
-		}
-		entries = append(entries, &entry)
-	}
-
-	return entries, nil
-}
-
-func (s *postgresStore) GetNumberOfTransports(context.Context) (map[types.Type]int, error) {
-	var tpRecords []Transport
-	response := map[types.Type]int{
-		types.STCP:  0,
-		types.STCPR: 0,
-		types.SUDPH: 0,
-		types.DMSG:  0,
-	}
-	if err := s.client.Find(&tpRecords).Error; err != nil {
-		return response, err
-	}
-	for _, record := range tpRecords {
-		response[types.Type(record.Type)]++
-	}
-	return response, nil
-}
-
-func (s *postgresStore) GetAllTransports(_ context.Context, selfTransports bool) ([]*transport.Entry, error) {
-	var tpRecords []Transport
-	if err := s.client.Find(&tpRecords).Error; err != nil {
-		return nil, ErrTransportNotFound
-	}
-
-	var entries []*transport.Entry
-
-	for _, tpRecord := range tpRecords {
-		entry, err := makeEntry(tpRecord)
-		if err != nil {
-			return nil, err
-		}
-		if !selfTransports {
-			if entry.Edges[0] == entry.Edges[1] {
-				continue
-			}
-		}
-		entries = append(entries, &entry)
-	}
-
-	return entries, nil
-}
-
 func (s *postgresStore) Close() {
 	close(s.closeC)
-}
-
-func makeEntry(record Transport) (transport.Entry, error) {
-	cipher1 := cipher.PubKey{}
-	if err := cipher1.UnmarshalText([]byte(record.EdgeA)); err != nil {
-		return transport.Entry{}, err
-	}
-
-	cipher2 := cipher.PubKey{}
-	if err := cipher2.UnmarshalText([]byte(record.EdgeB)); err != nil {
-		return transport.Entry{}, err
-	}
-
-	entry := transport.Entry{}
-	entry.Label = transport.Label(record.Label)
-	entry.Type = types.Type(record.Type)
-	entry.ID = uuid.MustParse(record.TransportID)
-	entry.Edges = [2]cipher.PubKey{cipher1, cipher2}
-
-	return entry, nil
-}
-
-// Transport is model (structure) for transports table
-type Transport struct { //TODO (mohammed): good to use transport.Entry model here
-	ID          uint `gorm:"primarykey"`
-	CreatedAt   time.Time
-	TransportID string `gorm:"unique"`
-	EdgeA       string
-	EdgeB       string
-	Type        string
-	Label       string
 }
